+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/model/metamodel/AnnotationPropertyAttributes.java
@@ -39,6 +39,6 @@ class AnnotationPropertyAttributes extends PropertyAttributes {
         this.simpleLiteral = oap.simpleLiteral();
         this.datatype = typeBuilderContext.resolveNamespace(oap.datatype());
         this.language = resolveLanguage(fieldValueCls);
-        validator.validateAnnotationPropertyField(field, oap);
+        validator.validateLiteralFieldMapping(field, this);
     }
 }