+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/model/metamodel/DataPropertyAttributesTest.java
@@ -1,11 +1,11 @@
 /**
  * Copyright (C) 2022 Czech Technical University in Prague
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any
  * later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
@@ -48,9 +48,7 @@ class DataPropertyAttributesTest {
     void resolveInvokesDataPropertyFieldValidation() throws Exception {
         final DataPropertyAttributes sut = initSystemUnderTest();
         sut.resolve(OWLClassA.getStrAttField(), metamodelBuilder, OWLClassA.getStrAttField().getType());
-        verify(validator)
-                .validateDataPropertyField(OWLClassA.getStrAttField(), OWLClassA.getStrAttField().getAnnotation(
-                        OWLDataProperty.class));
+        verify(validator).validateLiteralFieldMapping(OWLClassA.getStrAttField(), sut);
     }
 
     private DataPropertyAttributes initSystemUnderTest() {