+++ jopa-integration-tests/src/main/java/cz/cvut/kbss/jopa/test/OWLClassM.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2022 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.test;
 
@@ -19,6 +17,7 @@ import cz.cvut.kbss.jopa.model.annotations.OWLClass;
 import cz.cvut.kbss.jopa.model.annotations.OWLDataProperty;
 import cz.cvut.kbss.jopa.test.environment.Generators;
 import cz.cvut.kbss.jopa.vocabulary.XSD;
+import cz.cvut.kbss.ontodriver.model.LangString;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -69,6 +68,9 @@ public class OWLClassM {
 
     @OWLDataProperty(iri = Vocabulary.p_m_explicitDatatype, datatype = XSD.DURATION)
     private String explicitDatatype;
+
+    @OWLDataProperty(iri = Vocabulary.p_m_langString)
+    private LangString langString;
 
     public enum Severity {
         LOW, MEDIUM, HIGH
@@ -178,6 +180,14 @@ public class OWLClassM {
         this.explicitDatatype = explicitDatatype;
     }
 
+    public LangString getLangString() {
+        return langString;
+    }
+
+    public void setLangString(LangString langString) {
+        this.langString = langString;
+    }
+
     @Override
     public String toString() {
         return "OWLCLassM{" +
@@ -193,6 +203,7 @@ public class OWLClassM {
                 ", simpleLiteral=" + simpleLiteral +
                 ", stringCollection=" + stringCollection +
                 ", explicitDatatype=" + explicitDatatype +
+                ", langString=" + langString +
                 '}';
     }